(nearly) complete rewrite of KURL functionality. Now all template based to not repeat code fragments over and over again, and making sure that incoming strings are not re-encoded (with possibly different URL encode settings) if they were not changed. Introducing a class "URLEncoded", which knows about the difference of original encoding and change-forced encoding, for both strings and property sheets (key value maps). Adding table based encoders for the URL encoding that are one-time initialized at startup.
The problem with URL encoding is that it is not deterministic. Incoming
encodings can encode any non-ASCII character or only a few. The plus
sign (as an escape for space) can be used or not. In result, it is not
possible to reproduce a origin URL without keeping track of the
originally URL encoded string. The solution is to keep a reference to
the original (escaped) string unless the decoded string got changed by
the user, and output the original in that case.